requirements.txt
--extra-index-url https://azuremlsdktestpypi.azureedge.net/promptflow/
flask
requests
gunicorn
flask-cors
python-dotenv
promptflow
./whl/prompty-0.1.20240411.181111-py3-none-any.whl